+++ source/routes.js
@@ -1,30 +1,14 @@
 const express = require('express');
-const crypto = require('crypto');
-const connection = require('./database/connection');
-const bcrypt = require('bcrypt');
 
+
+const UserController = require('./Controllers/UserController');
+const SessionController = require('./Controllers/SessionController');
 const routes = express.Router();
 
-routes.get('/users', async (request, response) =>{
-    const users = await connection('users').select('*');
+routes.post('/sessions', SessionController.create)
 
-    return response.json(users);
-});
 
-routes.post('/users', async (request, response) => {
-    const {
-        username,
-        password
-    } = request.body;
-
-    const id = crypto.randomBytes(4).toString('HEX');
-    const hash = await bcrypt.hash(password, 10);
-    
-    console.log(connection);
-    
-    await connection('users').insert({username, id, password:hash});
-    
-    return response.json({id});
-});
+routes.get('/users', UserController.index);
+routes.post('/users', UserController.create);
 
 module.exports = routes;
